+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGELOG
 
+### 1.20.3 (2024-XX-XX)
+- **Feature:** Added new interface `isAppDeeplinkSupported` which helps to check if the application have whitelisted/allows the deeplink to open
+- **Feature:** Added new interface `canOpenAppDeeplink` which helps to check if the device can open the deeplink scheme
+- **Feature:** Added new interface `getPhoneNumber` to get the phone number from the native application
+- **Fix:** Updated `MAAnalyticsInfo` in SDK to have more parameters
+
 ### 1.20.2 (2024-03-27)
 - **Fix:** Remove dependecy of `libbase64` and we moved the encode/decode internally in SDK